requirements.txt
requests
jinja2
click
pyyaml
pyats